requirements.txt
anyio==4.6.2.post1
async-timeout==4.0.3
asyncpg==0.30.0
bcrypt==4.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
cryptography==44.0.0
decorator==5.1.1
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.95.2
fastapi-cli==0.0.5
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.8.2
joblib==1.4.2
markdown-it-py==3.0.0
mdurl==0.1.2
nodeenv==1.8.0
numpy==2.2.0
openai==1.58.1
pandas==2.2.3
passlib==1.7.4
pbr==6.1.0
platformdirs==3.11.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.18
Pygments==2.18.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-gnupg==0.5.2
python-jose==3.3.0
python-multipart==0.0.6
pytz==2024.2
PyYAML==6.0.2
requests==2.31.0
rich==13.9.4
rsa==4.9
scalar_fastapi==1.0.3
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlalchemy-migrate==0.13.0
sqlalchemy2-stubs==0.0.2a34
sqlparse==0.5.2
starlette==0.27.0
Tempita==0.6.0
threadpoolctl==3.5.0
tqdm==4.67.1
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
virtualenv==20.23.0
watchfiles==1.0.0
websockets==14.1
